Node transfer complete

+++ libsensor/source/load_count.c
@@ -74,7 +74,7 @@ static void load_close(ArrayList momentLoads, struct Node *client, uint32_t inte
 
 }
 
-void load_bytes_add(ArrayList momentLoads, int len) {
+void load_bytes_add(ArrayList momentLoads, size_t len) {
 	struct NodeLoad *mload = load_get_last(momentLoads);
 	if (!mload) {
 		mload = load_create_item(momentLoads);
